Python_Pizza
This is my project in the python programming language, it has a little functionality. This project contains a small pizzeria. In it, you can work as an administrator (password - 1234) or order pizza and drinks. You can also register in the data system by entering your name, age and password (it is not necessary to enter real data)